Combined sharable states into a single file to make them more managable
