+++ src/StateControllers/HyperFlushStateController.cpp
@@ -1,14 +1,10 @@
 #include <StateControllers/HyperFlushStateController.hpp>
-#include <States/Flush.hpp>
-#include <States/FlushVolume.hpp>
-#include <States/OffshootPreload.hpp>
-#include <States/Stop.hpp>
-#include <States/Idle.hpp>
+#include <States/Shared.hpp>
 
 void HyperFlushStateController::setup() {
 	using namespace HyperFlush;
-	registerState(Flush(), FLUSH, OFFSHOOT_PRELOAD);
-	registerState(OffshootPreload(), OFFSHOOT_PRELOAD, STOP);
-	registerState(Stop(), STOP, IDLE);
-	registerState(Idle(), IDLE);
+	registerState(SharedStates::Flush(), FLUSH, OFFSHOOT_PRELOAD);
+	registerState(SharedStates::OffshootPreload(), OFFSHOOT_PRELOAD, STOP);
+	registerState(SharedStates::Stop(), STOP, IDLE);
+	registerState(SharedStates::Idle(), IDLE);
 }
