+++ src/StateControllers/MainStateController.cpp
@@ -1,9 +1,6 @@
 #include <StateControllers/MainStateController.hpp>
 #include <Application/Application.hpp>
-#include <States/Flush.hpp>
-#include <States/Sample.hpp>
-#include <States/Dry.hpp>
-#include <States/Preserve.hpp>
+#include <States/Shared.hpp>
 
 void Main::Idle::enter(KPStateMachine & sm) {
 	Application & app = *static_cast<Application *>(sm.controller);
@@ -32,10 +29,10 @@ void MainStateController::setup() {
 	using namespace Main;
 
 	// FLUSH -> SAMPLE -> DRY -> PRESERVE -> STOP -> IDLE
-	registerState(Flush(), FLUSH, SAMPLE);
-	registerState(Sample(), SAMPLE, DRY);
-	registerState(Dry(), DRY, PRESERVE);
-	registerState(Preserve(), PRESERVE, STOP);
+	registerState(SharedStates::Flush(), FLUSH, SAMPLE);
+	registerState(SharedStates::Sample(), SAMPLE, DRY);
+	registerState(SharedStates::Dry(), DRY, PRESERVE);
+	registerState(SharedStates::Preserve(), PRESERVE, STOP);
 	registerState(Stop(), STOP, IDLE);
 	registerState(Idle(), IDLE);
 }
