--- src/States/Dry.hpp
@@ -1,8 +1,0 @@
-#pragma once
-#include <KPState.hpp>
-
-class Dry : public KPState {
-public:
-	unsigned long time = 10;
-	void enter(KPStateMachine & sm) override;
-};
