--- src/States/Idle.cpp
@@ -1,4 +1,0 @@
-#include <States/Idle.hpp>
-#include <Application/Application.hpp>
-
-void Idle::enter(KPStateMachine & sm) {}
