--- src/States/Idle.hpp
@@ -1,7 +1,0 @@
-#pragma once
-#include <KPState.hpp>
-
-class Idle : public KPState {
-public:
-	void enter(KPStateMachine & sm) override;
-};
