--- src/States/OffshootPreload.hpp
@@ -1,8 +1,0 @@
-#pragma once
-#include <KPState.hpp>
-
-class OffshootPreload : public KPState {
-public:
-	int preloadTime = 5;
-	void enter(KPStateMachine & sm) override;
-};
