--- src/States/Stop.hpp
@@ -1,7 +1,0 @@
-#pragma once
-#include <KPState.hpp>
-
-class Stop : public KPState {
-public:
-	void enter(KPStateMachine & sm) override;
-};
